Move core types into base.py and delete data_types.py
